backend/requirements.txt
fastapi>=0.110
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
alembic
pydantic>=2.0
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-cloud-storage
reportlab
httpx
pytest
pytest-asyncio
aiosqlite
